Update AHK script
The links to the AHK script were broken, so I updated them to the latest version. I also added a link to an explanation of its features.
Also, I changed the link to point at a release instead of the `main` branch of the nasin-nanpa repo. This way it won't break when the repo is rearranged.